Moved enginetest.go to new enginetest package, started breaking it into generic / memdb files.
